requirements.txt
django==2.2.24
pytz==2018.4
pillow==7.2
django-recaptcha==2.0.6
django-allauth==0.40.0
